+++ src/main/kotlin/racingcar/view/component/RacingResultComponent.kt
@@ -1,11 +1,11 @@
 package racingcar.view.component
 
-import racingcar.view.container.RoundContainer
+import racingcar.view.container.RoundListContainer
 import racingcar.view.container.WinnerContainer
 
 class RacingResultComponent : Component {
     override fun render() {
-        RoundContainer().render()
+        RoundListContainer().render()
         WinnerContainer().render()
     }
 }
